Set optional features at module level
