--- examples/__init__.py
@@ -1,1 +1,0 @@
-from . import boids, forest_fire, game_of_life, opinion_dynamics
