--- examples/boids/__init__.py
@@ -1,1 +1,0 @@
-from . import evolutionary, hard_coded, rl
