+++ src/esquilax/__init__.py
@@ -1,7 +1,7 @@
 """
 Esquilax package
 """
-from . import ml, reductions, transforms, typing, utils
+from . import reductions, transforms, typing, utils
 from .batch_runner import batch_sim_runner
 from .runner import sim_runner
 from .sim import Sim
