+++ src/esquilax/ml/evo/training.py
@@ -10,9 +10,10 @@ within the same loop.
 from __future__ import annotations
 
 from functools import partial
-from typing import TYPE_CHECKING, Collection, Optional, Tuple
+from typing import Collection, Optional, Tuple
 
 import chex
+import evosax
 import jax
 import jax.numpy as jnp
 import jax_tqdm
@@ -22,15 +23,10 @@ from esquilax.sim import Sim
 from esquilax.typing import TSimParams, TypedPyTree
 
 from . import tree_utils
-from ._import_check import requires_evosax
 from .strategy import Strategy
 from .types import TrainingData
 
-if TYPE_CHECKING:
-    import evosax
 
-
-@requires_evosax
 @partial(
     jax.jit,
     static_argnames=(
@@ -224,7 +220,6 @@ def train(
     return evo_states, rewards
 
 
-@requires_evosax
 @partial(
     jax.jit,
     static_argnames=(
